+++ flutter/test/default_integrations_test.dart
@@ -176,21 +176,8 @@ void main() {
         fixture.options.sdk.integrations.contains('nativeSdkIntegration'));
   });
 
-  test('loadContextsIntegration adds integration on ios', () async {
+  test('loadContextsIntegration adds integration', () async {
     _channel.setMockMethodCallHandler((MethodCall methodCall) async {});
-
-    final integration = loadContextsIntegration(fixture.options, _channel);
-
-    await integration(fixture.hub, fixture.options);
-
-    expect(true,
-        fixture.options.sdk.integrations.contains('loadContextsIntegration'));
-  });
-
-  test('loadContextsIntegration do not throw', () async {
-    _channel.setMockMethodCallHandler((MethodCall methodCall) async {
-      throw null;
-    });
 
     final integration = loadContextsIntegration(fixture.options, _channel);